requirements.txt
streamlit==1.31.0
sqlalchemy==2.0.25
pypdf==4.0.1
python-docx==1.1.0
reportlab==4.0.9
requests==2.31.0
plotly==5.18.0
